Frozen Food - Text Classification
In this notebook, we will try to categorize text from a mini survey about frozen food business owners' preference for Fulfillment Center.

A short brief about fulfillment center is a business warehouse that takes care of other person's product. Whenever there are orders from e-commerce or affiliated party to the product's owner, the fulfillment center will act as the representative of the business owner and will send the items instead of the owner. This Fulfillment Center business will definitely cut the operational cost of the business owner.

This notebook will try to analyze the survey result from 105 respondents of Frozen Food Business owners, about their preference for using Fulfillment Center in their business.