ATM Action (Class)
The goal of this project is to make an actionable ATM action by utilizing Class in Python. Class is like a blueprint for making an Object Oriented Programming, consisting of a variety of Methods (def Function)

def init serves as the global function for variables that are used inside the class, self is used to personalized the variables.

In this section, we will build the class called State. This will be used to create an OOP of the ATM function. The password for the ATM itself is called "hacker". If the password is wrong, any of the command will become unauthorized, hence the output will be authorized and the action will be execute.